move reflection probe to legacy
