+++ cocos/rendering/index.ts
@@ -39,6 +39,3 @@ export { PipelineStateManager } from './pipeline-state-manager';
 
 export { PipelineEventProcessor, PipelineEventType } from './pipeline-event';
 export { DebugView } from './debug-view';
-
-export { ReflectionProbeFlow } from './reflection-probe/reflection-probe-flow';
-export { ReflectionProbeStage } from './reflection-probe/reflection-probe-stage';
